refactor: DDD 아키텍처 적용 및 환경 설정 정리 (#1)
