+++ src/main/kotlin/infrastructure/config/HttpConfig.kt
@@ -1,0 +1,22 @@
+package infrastructure.config
+
+import io.ktor.http.*
+import io.ktor.server.application.*
+import io.ktor.server.plugins.cors.routing.*
+
+fun Application.configureHTTP() {
+    install(CORS) {
+        allowMethod(HttpMethod.Options)
+        allowMethod(HttpMethod.Put)
+        allowMethod(HttpMethod.Delete)
+        allowMethod(HttpMethod.Patch)
+        allowMethod(HttpMethod.Get)
+        allowMethod(HttpMethod.Post)
+        allowHeader(HttpHeaders.Authorization)
+        allowHeader(HttpHeaders.ContentType)
+        allowHeader("X-Requested-With")
+        
+        // 개발 환경에서는 모든 호스트 허용, 운영에서는 제한 필요
+        anyHost() // TODO: 운영 환경에서는 특정 도메인만 허용
+    }
+}
